+++ backend/cmd/main.go
@@ -3,10 +3,10 @@ package main
 import (
 	"github.com/gin-gonic/gin"
 	"github.com/isiyar/daily-energy/backend/config"
+	"github.com/isiyar/daily-energy/backend/internal/adapters/db"
+	"github.com/isiyar/daily-energy/backend/internal/adapters/http/router"
+	"github.com/isiyar/daily-energy/backend/internal/adapters/repository"
 	"github.com/isiyar/daily-energy/backend/internal/app/usecase"
-	"github.com/isiyar/daily-energy/backend/internal/infrastructure/db"
-	"github.com/isiyar/daily-energy/backend/internal/infrastructure/http/router"
-	"github.com/isiyar/daily-energy/backend/internal/infrastructure/repository"
 	"github.com/isiyar/daily-energy/backend/internal/interfaces/http/handler"
 )
